+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+---
+
+## [0.1.43] - 2026-09-14
+
 ### Performance
 
 - **The upload and backup handlers no longer block the async executor on